Tailwindcss Installation + Upgrade Ruby BREAKING
