+++ .gitignore
@@ -35,3 +35,7 @@
 /config/master.key
 
 .env
+/app/assets/builds/*
+!/app/assets/builds/.keep
+
+.DS_Store
