+++ Procfile.dev
@@ -1,0 +1,2 @@
+web: bin/rails server -b=0.0.0.0
+css: bin/rails tailwindcss:watch
