+++ app/assets/stylesheets/application.css
@@ -13,3 +13,4 @@
  *= require_tree .
  *= require_self
  */
+ @import "tailwindcss";
