+++ app/assets/tailwind/application.css
@@ -1,0 +1,1 @@
+@import "tailwindcss";
