+++ app/views/member/index.html.erb
@@ -1,14 +1,16 @@
-<h1>Welcome, Member!</h1>
-<p>We are glad to have you here.</p>
+<h1 class="text-3xl font-bold mb-4">Welcome, Member!</h1>
+<p class="text-lg mb-6">We are glad to have you here.</p>
 
 <% if current_member.role >= 5 %>
-    <%= link_to 'Admin Dashboard', admin_path, class: 'btn btn-primary' %>
+  <%= link_to 'Admin Dashboard', admin_path, class: 'btn btn-primary bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mb-4' %>
 <% end %>
 
 <% if current_member.role >= 1 %>
-    <%= link_to 'Guest Speakers', speakers_path, class: 'btn btn-primary' %>
+  <%= link_to 'Guest Speakers', speakers_path, class: 'btn btn-primary bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mb-4' %>
 <% end %>
 
-<%= render partial: 'shared/events_list' %>
+<div class="mb-6">
+  <%= render partial: 'shared/events_list' %>
+</div>
 
-<%= link_to 'Logout', destroy_member_session_path, method: :delete, class: 'btn btn-primary' %>
+<%= link_to 'Logout', destroy_member_session_path, method: :delete, class: 'btn btn-primary bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded' %>
