--- app/views/member/sessions/new.html.erb
@@ -1,4 +1,0 @@
-<h2>You're Logged out! Log in to the Dashboard!!</h2>
-<%= button_to member_google_oauth2_omniauth_authorize_path, class: "btn btn-danger", method: :post, data: {turbo: "false"} do %>
-<i class="fab fa-google mobile-text"></i><span class = "mobile-font">&nbsp;Google</span>
-<% end %>
